fix(countdown): Separate timeUtils & countdownUtils
