+++ src/index.ts
@@ -26,7 +26,7 @@ import Spinner, {SpinnerProps} from "./spinner/Spinner";
 import Tab, {TabItem, TabProps} from "./tab/Tab";
 import Countdown, {CountdownItem, CountdownProps} from "./countdown/Countdown";
 import useCountDownTimer from "./core/utils/hooks/useCountdownTimer";
-import Textarea, { TextareaProps } from "./form/textarea/Textarea";
+import Textarea, {TextareaProps} from "./form/textarea/Textarea";
 import Avatar, {AvatarProps} from "./avatar/Avatar";
 
 export {
